requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
beautifulsoup4
ollama
fastapi-limiter
redis
groq
